Fixed wrong header declaration

+++ include/CSVReader.h
@@ -6,5 +6,5 @@
 namespace CSV
 {
     typedef std::vector<std::string> Row;
-    std::vector<Row> Parse(const char* data, unsigned int length);
+    std::vector<Row> Parse(const char* data, size_t length);
 }
